Add "validation image class"

+++ handlers/profile_handler.php
@@ -1,44 +1,31 @@
 <?php
 if (isset($_POST['setAvatar'])) {
 
-  $filename = $_FILES['image']['name'];
-  $tempName = $_FILES['image']['tmp_name'];
-  $extension = strToLower(pathinfo($filename, PATHINFO_EXTENSION));
-
-  $errors = [];
-
-  if ($_FILES["image"]["error"]) {
-    $errors[] = 'File not found';
-  }
-
-  if ($_FILES["image"]["size"] > 5000000) {
-    $errors[] = "File is too large.";
-  }
-
-
-  if ($extension != "jpg" && $extension != "png" && $extension != "jpeg") {
-    $errors[] = 'Format of your file doesn\'t support';
-  }
+  $validationImage = new ValidationImage($_FILES);
+  $errors = $validationImage->validateFile();
 
   if (!$errors) {
-    $updatedFileName = $_SESSION['user_id'] . '.' . $extension;
+
+    $findAll = glob("../img/{$_SESSION['user_id']}*");
+
+    foreach ($findAll as $img) {
+      unlink($img);
+    }
+
+    $updatedFileName = $_SESSION['user_id'] . '.' . $validationImage->getExtension();
     $path = "../img/" . $updatedFileName;
 
     $sql = 'UPDATE users SET img = :img WHERE user_id = :user_id';
     $stmt = $pdo->prepare($sql);
     $stmt->execute(['img' => $updatedFileName, 'user_id' => $_SESSION['user_id']]);
 
-    if (file_exists($path)) {
-      unlink($path);
-    }
 
-    if (move_uploaded_file($tempName, $path)) {
+    if (move_uploaded_file($validationImage->getTempName(), $path)) {
       header("Location: " . $_SERVER['PHP_SELF']);
       die();
     }
   }
 
-  echo 'Image doesn\'t uploaded<br>';
   foreach ($errors as $error) {
     echo $error . '<br>';
   }
@@ -46,10 +33,11 @@ if (isset($_POST['setAvatar'])) {
 }
 
 
-if(isset($_POST['deleteAvatar'])) {
-  $sql = 'UPDATE users SET img = null WHERE user_id = :user_id';
-  $stmt = $pdo->prepare($sql);
-  $stmt->execute(['user_id' => $_SESSION['user_id']]);
-  unlink($_POST['fileName']);
+if (isset($_POST['deleteAvatar'])) {
+  if ($_POST['fileName'] != '../img/default.png') {
+    $sql = 'UPDATE users SET img = null WHERE user_id = :user_id';
+    $stmt = $pdo->prepare($sql);
+    $stmt->execute(['user_id' => $_SESSION['user_id']]);
+    unlink($_POST['fileName']);
+  }
 }
-
